skills/L1_Streamlit/pyproject.toml
[project]
name = "l1-commentary"
version = "4.0.0"
description = "L1 Financial Commentary Generator for Product Categories"
requires-python = ">=3.9"
dependencies = [
    "snowflake-connector-python>=3.0.0",
]
